requirements.txt
# ExtratorADK - Extração de Documentos Brasileiros (RG, CNH, CPF)

# Google ADK e Gemini
google-adk==1.15.1
google-generativeai>=0.8.3

# Web server (necessário para ADK web)
python-multipart>=0.0.9
aiohttp>=3.9.0

# Processamento de imagens
Pillow>=11.1.0

# Validação de documentos brasileiros
validate-docbr>=1.10.0

# Utilitários
python-dotenv==1.0.1
python-dateutil>=2.9.0

# Logging
loguru==0.7.3

# Testes
pytest==8.3.4
pytest-cov==6.0.0
